refactor: style 파일명 네이밍 컨벤션에 맞게 수정 (#21)
